+++ src/code42cli/profile/password.py
@@ -1,13 +1,16 @@
+from __future__ import print_function
 import keyring
 from getpass import getpass
-import c42sec.profile._config as config
-from c42sec.profile._config import ConfigurationKeys
+
+import code42cli.profile.config as config
+from code42cli.profile.config import ConfigurationKeys
 
 
-_ROOT_SERVICE_NAME = u"c42sec"
+_ROOT_SERVICE_NAME = u"code42cli"
 
 
 def get_password(prompt_if_not_exists=True):
+    """Gets your currently stored password for your username / authority URL combo."""
     profile = config.get_config_profile()
     service_name = _get_service_name(profile)
     username = _get_username(profile)
@@ -19,6 +22,7 @@ def get_password(prompt_if_not_exists=True):
 
 
 def set_password():
+    """Prompts and sets your password for your username / authority URL combo."""
     password = getpass()
     profile = config.get_config_profile()
     service_name = _get_service_name(profile)