+++ tests/profile/test_profile.py
@@ -1,48 +1,48 @@
 import pytest
 from argparse import ArgumentParser
-from c42sec.profile import profile
+from code42cli.profile import profile
 
 
 @pytest.fixture
 def username_setter(mocker):
-    return mocker.patch("c42sec.profile._config.set_username")
+    return mocker.patch("code42cli.profile.config.set_username")
 
 
 @pytest.fixture
 def mark_as_set_function(mocker):
-    return mocker.patch("c42sec.profile._config.mark_as_set")
+    return mocker.patch("code42cli.profile.config.mark_as_set")
 
 
 @pytest.fixture
 def authority_url_setter(mocker):
-    return mocker.patch("c42sec.profile._config.set_authority_url")
+    return mocker.patch("code42cli.profile.config.set_authority_url")
 
 
 @pytest.fixture
 def ignore_ssl_errors_setter(mocker):
-    return mocker.patch("c42sec.profile._config.set_ignore_ssl_errors")
+    return mocker.patch("code42cli.profile.config.set_ignore_ssl_errors")
 
 
 @pytest.fixture
 def password_setter(mocker):
-    return mocker.patch("c42sec.profile._password.set_password")
+    return mocker.patch("code42cli.profile.password.set_password")
 
 
 @pytest.fixture
 def password_getter(mocker):
-    return mocker.patch("c42sec.profile._password.get_password")
+    return mocker.patch("code42cli.profile.password.get_password")
 
 
 @pytest.fixture
 def profile_not_set_state(mocker):
-    profile_verifier = mocker.patch("c42sec.profile._config.profile_has_been_set")
+    profile_verifier = mocker.patch("code42cli.profile.config.profile_has_been_set")
     profile_verifier.return_value = False
     return profile_verifier
 
 
 @pytest.fixture
 def profile_is_set_state(mocker):
-    profile_verifier = mocker.patch("c42sec.profile._config.profile_has_been_set")
+    profile_verifier = mocker.patch("code42cli.profile.config.profile_has_been_set")
     profile_verifier.return_value = True
     return profile_verifier
 
@@ -64,17 +64,15 @@ def test_init_adds_parser_that_can_parse_set_command(config_parser):
     subcommand_parser = ArgumentParser().add_subparsers()
     profile.init(subcommand_parser)
     profile_parser = subcommand_parser.choices.get("profile")
-
-    # Commands that require a value will fail here if not provided
-    assert profile_parser.parse_args(
+    profile_parser.parse_args(
         ["set", "-s", "server-arg", "-p", "-u", "username-arg", "--enable-ssl-errors"]
     )
 
 
-def test_get_profile_returns_object_from_config_file(config_parser, config_profile):
+def test_get_profile_returns_object_from_config_profile(config_parser, config_profile):
     user = profile.get_profile()
 
-    # Values from config_file fixture
+    # Values from config_profile fixture
     assert (
         user.username == "test.username"
         and user.authority_url == "https://authority.example.com"