+++ tests/securitydata/conftest.py
@@ -1,0 +1,2 @@
+ROOT_PATH = "code42cli.securitydata"
+SUBCOMMANDS_PATH = "{0}.subcommands".format(ROOT_PATH)