+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
@@ -10,22 +11,36 @@ how a consumer would use the library (e.g. adding unit tests, updating documenta
 ## Unreleased
 
 ### Removed
-- Removed config file settings and `-c` CLI arg. Use `c42sec profile set`.
-- Removed `--clear-password` CLI argument. Use `c42sec profile set -p`. You will be prompted.
+
+- Removed config file settings and `-c` CLI arg. Use `code42 profile set`.
+- Removed `--clear-password` CLI argument. Use `code42 profile set -p`. You will be prompted.
+- Removed top-level destination args. Use subcommands `write-to`. `send-to`, `print` off of `code42 security data`.
 
 ### Added
-- Added ability to view your profile: `c42sec profile show`.
+
+- Added ability to view your profile: `code42 profile show`.
+- Added `securitydata` subcommands:
+    - Use `code42 securitydata write-to` to output to a file.
+    - Use `code42 securitydata send-to` to output to a server.
+    - Use `code42 securitydata print` to outputs to stdout.
+    - Use `code42 securitydata clear-cursor` to remove the stored cursor for 'incremental' mode.
+- Added support for raw JSON queries via `code42 securitydata [subcommand] --advanced-query [JSON]`.
 
 ### Changed
-- Renamed `c42aed` to `c42sec`.
-- Moved CLI arguments `-s`, `-u`, and `--ignore-ssl-errors` to `c42sec profile set` command.
 
+- Renamed base command `c42aed` to `code42`.
+- Moved CLI arguments `-s`, `-u`, and `--ignore-ssl-errors` to `code42 profile set` command.
+- Renamed and moved top-level `-r` flag. 
+    - Use `-i` on one of these `securitydata` subcommands `write-to`. `send-to`, `print`.
+- Moved search arguments to individual `securitydata` subcommands `write-to`. `send-to`, `print`.
 
 ## 0.1.1 - 2019-10-29
 
 ### Fixed
+
 - Issue where IOError message was inaccurate when using the wrong port for server destinations.
 
 ### Added
+
 - Error handling for all socket errors.
 - Error handling for IOError 'connection refused'.